Updated Makefile Updated main.c
