+++ .gitignore
@@ -51,5 +51,5 @@ Module.symvers
 Mkfile.old
 dkms.conf
 
-# Linux executable
-ccalc
+obj
+bin
