+++ Makefile
@@ -1,5 +1,53 @@
-dir_source := source
+WORKDIR = `pwd`
 
+CC = gcc
+CXX = g++
+AR = ar
+LD = g++
+WINDRES = 
 
-program: $(dir_source)/main.c
-	gcc $(dir_source)/main.c -o ccalc
+INC = 
+CFLAGS = -Wall
+RESINC = 
+LIBDIR = 
+LIB = 
+LDFLAGS = 
+
+INC_RELEASE = $(INC)
+CFLAGS_RELEASE = $(CFLAGS) -O2
+RESINC_RELEASE = $(RESINC)
+RCFLAGS_RELEASE = $(RCFLAGS)
+LIBDIR_RELEASE = $(LIBDIR)
+LIB_RELEASE = $(LIB) 
+LDFLAGS_RELEASE = $(LDFLAGS) -s
+OBJDIR_RELEASE = obj/Release
+DEP_RELEASE = 
+OUT_RELEASE = bin/Release/Tech
+
+OBJ_RELEASE = $(OBJDIR_RELEASE)/main.o
+
+all: release
+
+clean: clean_release
+
+before_release: 
+	test -d bin/Release || mkdir -p bin/Release
+	test -d $(OBJDIR_RELEASE) || mkdir -p $(OBJDIR_RELEASE)
+
+after_release: 
+
+release: before_release out_release after_release
+
+out_release: before_release $(OBJ_RELEASE) $(DEP_RELEASE)
+	$(LD) $(LIBDIR_RELEASE) -o $(OUT_RELEASE) $(OBJ_RELEASE)  $(LDFLAGS_RELEASE) $(LIB_RELEASE)
+
+$(OBJDIR_RELEASE)/main.o: main.c
+	$(CC) $(CFLAGS_RELEASE) $(INC_RELEASE) -c main.c -o $(OBJDIR_RELEASE)/main.o
+
+clean_release: 
+	rm -f $(OBJ_RELEASE) $(OUT_RELEASE)
+	rm -rf bin/Release
+	rm -rf $(OBJDIR_RELEASE)
+
+.PHONY: before_release after_release clean_release
+
